Make failures about missing/unparseable NAME sections clearer
Currently, if there is no `=head1 NAME` in the main module,
the following warning is emitted:

  Use of uninitialized value in pattern match (m//) at <prefix>/App/ModuleBuildTiny.pm line <no>

This commit changes this to be a more obvious warning, so that in any
instances of:

1. If there is no =head1 NAME section.

2. =head1 NAME is empty.

3. =head1 NAME can't be parsed.

A warning indicating that its absence/unparsability is emitted.

meta.abstract is then still either 'undef' or an empty string, the same
as it presently was, but with clearer warnings.